refactor: Migrate import generator to typescript (LPS-119618)
